// apps/rawg-front-end/lib/api/model/game.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RAWG Video Games Database API
 * 
The largest open video games database.

### Why build on RAWG
- More than 350,000 games for 50 platforms including mobiles.
- Rich metadata: tags, genres, developers, publishers, individual creators, official websites, release dates,
Metacritic ratings.
- Where to buy: links to digital distribution services
- Similar games based on visual similarity.
- Player activity data: Steam average playtime and RAWG player counts and ratings.
- Actively developing and constantly getting better by user contribution and our algorithms.

### Terms of Use
- Free for personal use as long as you attribute RAWG as the source of the data and/or images and add an active
hyperlink from every page where the data of RAWG is used.
- Free for commercial use for startups and hobby projects with not more than 100,000 monthly active users or 500,000
page views per month. If your project is larger than that, email us at [api@rawg.io](mailto:api@rawg.io) for
commercial terms.
- No cloning. It would not be cool if you used our API to launch a clone of RAWG. We know it is not always easy
to say what is a duplicate and what isn't. Drop us a line at [api@rawg.io](mailto:api@rawg.io) if you are in doubt,
and we will talk it through.
- You must include an API key with every request. The key can be obtained at https://rawg.io/apidocs.
If you don’t provide it, we may ban your requests.

__[Read more](https://rawg.io/apidocs)__.

 * OpenAPI spec version: v1.0
 */
import type { GameRatings } from './gameRatings';
import type { GameAddedByStatus } from './gameAddedByStatus';
import type { GameEsrbRating } from './gameEsrbRating';
import type { GamePlatformsItem } from './gamePlatformsItem';

export interface Game {
  readonly id?: number;
  /**
   * @minLength 1
   * @pattern ^[-a-zA-Z0-9_]+$
   */
  readonly slug?: string;
  /** @minLength 1 */
  readonly name?: string;
  readonly released?: string;
  readonly tba?: boolean;
  readonly background_image?: string;
  rating: number;
  readonly rating_top?: number;
  readonly ratings?: GameRatings;
  readonly ratings_count?: number;
  readonly reviews_text_count?: string;
  readonly added?: number;
  readonly added_by_status?: GameAddedByStatus;
  readonly metacritic?: number;
  /** in hours */
  readonly playtime?: number;
  readonly suggestions_count?: number;
  readonly updated?: string;
  /** @nullable */
  esrb_rating?: GameEsrbRating;
  platforms?: GamePlatformsItem[];
}